+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.1.4: 2025-12-07
 
 * Add automatic detection and resolution of incomplete merges from interrupted script runs
+* Improve merge conflict handling with options to resolve, abort, or exit
 
 ### 1.1.3: 2025-11-15